(PC-7399) Refacto: create a reusable svgheader
