+++ src/features/bookings/pages/Bookings.tsx
@@ -3,6 +3,7 @@ import React from 'react'
 import styled from 'styled-components/native'
 
 import { _ } from 'libs/i18n'
+import SvgPageHeader from 'ui/components/headers/SvgPageHeader'
 import { Spacer, Typo } from 'ui/theme'
 
 import { useBookings } from '../api'
@@ -11,11 +12,14 @@ export const Bookings: React.FC = () => {
   const { data } = useBookings()
 
   return (
-    <Container>
-      <Spacer.Flex />
-      <Typo.Hero>{_(t`Bookings`)}</Typo.Hero>
-      <Spacer.Flex />
-    </Container>
+    <React.Fragment>
+      <SvgPageHeader title="Mes réservations" />
+      <Container>
+        <Spacer.Flex />
+        <Typo.Hero>{_(t`Bookings`)}</Typo.Hero>
+        <Spacer.Flex />
+      </Container>
+    </React.Fragment>
   )
 }
